Avoid using params and touch files for what should be output files
@corneliusroemer wrote in Slack¹:

> I’m using params because I couldn’t figure out how to get a list of
> outputs using expand with the variety of wildcards, this may be possible
> but params did it more easily

I think one issue he might have encountered was due to not shell quoting
the literal "{gene}" that's part of the --output-translations value.
Without quoting, the shell interprets it as an expansion and Nextclade
sees "translation_gene" instead of "translation_{gene}".

This commit reverts a number of changes in "Add translations as extra
ingest output" (fa5a131) that were only necessary due to the use of
params and touch files, thus making the diff across the branch's merge
base and tip smaller and simpler.

¹ <https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1686877758714089?thread_ts=1686858096.401379&cid=C01LCTT7JNN>