transform-rki: update `COLUMN_MAP`
The reorg of the RKI repo included updating the headers and columns of
the metadata file.

This commit updates the `COLUMN_MAP` to the new column headers.
It also drops two columns that are no longer available in the new
metadata TSV: `date_released` and `strain_gisaid`. These two columns are
not used in the transform pipeline and are not included in the output
TSV so this change should not affect the final output metadata TSV of
the open ncov-ingest pipeline.