Remove unused `run-nextclade-full` scripts and GH Action workflows
These have been unused since we transitioned to using touchfiles
to do full Nextclade runs.¹ These scripts are also outdated since we've
added new translation outputs to the automated workflow runs.²

This was also motivated by our recent push to deduplicate ingest
scripts. Since a lot of changes in `notify-on-job-start` unique to this
repo were added to support the `run-nextclade-full` scripts³, it will
simplify the centralized version of `notify-on-job-start` by removing
the unused scripts here.

¹ https://github.com/nextstrain/ncov-ingest/pull/380
² https://github.com/nextstrain/ncov-ingest/pull/404
³ https://github.com/nextstrain/ncov-ingest/compare/88be153e02acc5f318c970dcf44e8745128d469f...e898400a0ebf599dcb9e69f1f93dc8c108c97796