slack_notifications: Simplify DAG
Add the `notify_on_record_change` output as an input to the `notify_gisaid`
and `notify_genbank` rules to simplify the DAG.

Ensures the "notify.done" flag file signals the completion of all
notification rules. This will make it easier to track when the upload
rules can run in the following commit.