upload: Include notification touch file as input
Ensure that uploads only run after the notifications rules have run.
This prevents the race condition between diffs and uploads described
in https://github.com/nextstrain/ncov-ingest/issues/423.

This will slightly delay the uploads of files but is necessary to
support the diffs of local files and files on S3.