Remove unused variables and refactor `GENE_LIST`
Noted in previous PRs that that the `GENES` and `GENES_SPACE_DELIMITED`
variables are not needed¹ or used in the workflow,² so refactor the
`GENE_LIST` to be a hardcoded list of genes.

If we want to ensure that we do not miss any genes from the Nextclade
dataset, we could parse out the gene names from the dataset's
genome_annotation.gff file. However, I think that will over-complicate
the Snakemake workflow so I'm leaving the hardcoded list.

¹ https://github.com/nextstrain/ncov-ingest/pull/372#discussion_r1046020969
² https://github.com/nextstrain/ncov-ingest/pull/435#discussion_r1496332575