fetch-and-ingest: Add `trial_name` input
Update the fetch-and-ingest workflows to use the `trial_name` input to
determine whether the workflow should trigger downstream workflows
and the S3 destination prefix for uploaded files.

This changes the S3 destinations for both GISAID and GenBank test builds.

GISAID
  Old: s3://nextstrain-ncov-private/branch/${GITHUB_BRANCH}/
  New: s3://nextstrain-ncov-private/trial/${TRIAL_NAME}/

GenBank:
  Old: s3://nextstrain-staging/files/ncov/open/branch/${GITHUB_BRANCH}
  New: s3://nextstrain-data/files/ncov/open/trial/${TRIAL_NAME}

This follows the prefix patterns used by other more recent GitHub Action
workflows. By using the trial input, we also no longer have to maintain
extra copies of the "*-branch" workflows.

Note that I'm not updating the ingest-* workflows because they will be
made obsolete in the following commit.