fetch-and-ingest: Add `fetch_from_database` input
Update the fetch-and-ingest workflows to use the `fetch_from_database`
input to determine whether the workflow should fetch from upstream
databases. This means we no longer have to maintain extra copies of the
"ingest-*" workflows.

One downside is this input is _not_ included with the repository_dispatch
event. However, it seems like everyone's using the workflow_dispatch via
the GitHub UI instead of the repository_dispatch anyways:

```
$ gh api /repos/nextstrain/ncov-ingest/actions/runs?event=workflow_dispatch --jq .total_count --paginate
60
$ gh api /repos/nextstrain/ncov-ingest/actions/runs?event=repository_dispatch --jq .total_count --paginate
0
```