fetch-and-ingest: refactor to use pathogen-repo-build
Allows the GH Action workflows to stay connected to the AWS Batch jobs
when using the latest pathogen-repo-build workflow. This way, the status
of the GH Action workflow accurately reflects the status of the ingest
workflow.

We can then also use the Nextstrain pathogen workflow status page¹ to
monitor the success/failure of the automated workflows.

¹ https://nextstrain.github.io/status/pathogen-workflows.html