Bump CPUs for fetch-and-ingest workflows
I've been only bumping the memory but not the CPUs for the
fetch-and-ingest workflows. Might as well use all the compute that we
are paying for. GenBank should be using c5.9xlarge and GISAID should be
using c5.12xlarge, so bumping CPUs to match the instances.¹

Maybe this will magically help https://github.com/nextstrain/ncov-ingest/issues/446?

¹ <https://aws.amazon.com/ec2/instance-types/c5/>