fetch-and-ingest: Add Snakemake `--stats` output
Adding as part of #240 to help collect more data for tackling #446.

One unexpected behavior that I ran into when testing the `--stats`
option is that Snakemake doesn't generate the stats file if the
workflow exits with an error at any step.

Note that the Snakemake `--stats` option is not available starting with
Snakemake v8, so this will need to be removed when we eventually
upgrade Snakemake in our runtimes.