Move downloaded Nextclade executable to `data/nextclade`
Avoids clash of downloaded Nextclade executable with the Nextclade
command available in the environment.

Includes the side-effect of the downloaded executable being removed
as part of `bin/clean` when running the workflow without the
`keep_all_files=True` config param. This ensures that the workflow will
start from a clean slate.