fetch-cache-version: Use the Nextclade version JSON
Now that we are creating a Nextclade version JSON for the Nextclade TSV,
we can use it to fetch the cache version.

If the Nextclade TSV sha256sum doesn't match the value in the version
JSON, then we ignore the version JSON which results in the workflow
ignoring the Nextclade cache.