fix: make legacy mapping use `{clade} (Omicron)` by default
Resolves #471
I forgot to update the clade-legacy-mapping.yml with the last clade update (24C). As a result, the Nextstrain_clade output for 24C was wrong (`?` instead of `24C (Omicron)`).

To reduce maintenance effort, we now use `{value} (Omicron)` as the default mapping. This means one less file to update when clades are updated. If there's ever a non-Omicron again, we'd have to updated `clade-legacy-mapping.yml` but WHO has stopped giving variant names and there's nothing left that's not Omicron. So this should work into the indefinite future (until Emma stops using `Nextstrain_clade` - she's possibly the only user of that backwards-compat column)