Concatenate accession link files
Concatenates the downloaded accession links file to the existing
`source-data/accessions.tsv.gz` since the downloaded file seems to be
missing some of the existing links.¹

The transform scripts apply the accession links in order, so the
last matching accession link is used in the final metadata. This allows
us to default to the downloaded file which is the latest data.

¹ <https://github.com/nextstrain/ncov-ingest/pull/485#issuecomment-2549744920>