StandardizeDataRki: filter for "PANGOLIN_LATEST"
Based on <https://github.com/nextstrain/ncov-ingest/pull/476#discussion_r1780889881>,
we can filter for `PANGOLIN_LATEST` to get the latest lineage assignment.

If none of the lineages are marked as `PANGOLIN_LATEST`, then just use
the first one in the list since this was the behavior before the change.
If there are multiple `PANGOLIN_LATEST` lineages, then just use the
first one and output a warning. I've removed the assertion because this
should not block the whole ncov-ingest workflow.

Resolves <https://github.com/nextstrain/ncov-ingest/issues/478>