ParseGeographicColumnsGenbank: replace assertion with warning log
Prevent badly formatted data from causing the entire workflow to fail. 
Clearly, this hasn't been an issue since the `NameError` bug fixed in the 
previous commit has existed since it was implemented ~4 years ago and hasn't 
come up in our automated runs. However, it is annoying for the long running job 
to fail on a single bad data entry.