Add rules to use manual downloads from GISAID
Support ingest of manual downloads from the "Input to Augur pipeline" of EpiCoV.
Expects the manual downloads to be available locally as 
    - data/gisaid/YYYY-MM-DD-N-metadata.tsv
    - data/gisaid/YYYY-MM-DD-N-sequences.fasta

`YYYY-MM-DD` should represent the date of download and `N` should be the number 
of downloads. This allows for multiple downloads per day since there is a 
5,000 record limit per download. Specify the config param `gisaid_pairs` to 
be explicit about which pairs to use. Otherwise, the pipeline will automatically
use _all_ files that match the expected filename format. 

The metadata and sequences are linked and transformed to NDJSON in the format 
that matches our previous downloaded NDJSON cache from GISAID to require 
minimum updates to our transform pipeline. Records are deduplicated by the 
specified `gisaid_id_field` to guard against duplicate records from multiple 
downloads.