Add rules for ingesting unprocessed files and cache from S3
Add rules for ingesting the unprocessed GISAID files and the existing cache 
NDJSON on S3. Unprocessed GISAID files are downloaded and prepended to 
the cache so that the new records are prioritized during deduplication.

Once the updated cache has been uploaded to S3, then the unprocessed files are 
moved to processed on S3 so that they are not ingested in the next workflow run.